+++ src/countdown/__init__.py
@@ -11,7 +11,7 @@ __author__ = "Tanner B. Corcoran"
 __email__ = "tannerbcorcoran@gmail.com"
 __license__ = "MIT License"
 __copyright__ = "Copyright (c) 2023-present Tanner B. Corcoran"
-__version__ = "0.0.2"
+__version__ = "0.0.3"
 __description__ = ("An easy-to-use and highly customizable formatter used to turn a time value "
                    "into a countdown string")
 __url__ = "https://github.com/tanrbobanr/countdown"
@@ -19,10 +19,12 @@ __download_url__ = "https://pypi.org/project/countdown"
 
 __all__ = (
     "Countdown",
+    "TimeValue",
     "formatter",
     "constants"
 )
 
 from ._countdown import Countdown
+from .models import TimeValue
 from . import formatter
 from . import constants