+++ src/countdown/formatter.py
@@ -23,6 +23,7 @@ SOFTWARE.
 """
 
 from typing import Union
+from . import exceptions
 from . import constants
 from . import models
 from . import types
@@ -31,6 +32,8 @@ import logging
 import inspect
 import string
 import typing
+import prepr
+import re
 
 
 str_formatter = string.Formatter()
@@ -55,18 +58,72 @@ def _mangled_fmt(parent: str, child: str, _format_spec: str, _conversion: str) -
     return f"{{_{parent}__{child}{_conversion}{_format_spec}}}"
 
 
-class Flag:
+class AllPretty:
+    def __repr__(self, *args, **kwargs) -> prepr.pstr:
+        attrs = {k:getattr(self, k) for k in self.__slots__}
+        return prepr.prepr(self).kwargs(**attrs).build(simple=True, *args, **kwargs)
+
+
+class ParseArg(AllPretty):
+    __slots__ = ("key", "pretext", "fmt", "required")
+    def __init__(self, key: str, pretext: str, fmt: str, required: bool) -> None:
+        self.key = key
+        self.pretext = pretext
+        self.fmt = fmt
+        self.required = required
+
+
+class Flag(AllPretty):
     """Stores any valid flag in `VALID_FLAGS` along with any plural specifiers and extra defaults
     connected to it.
     
     """
     _log = plogging.setup_new("Flag", level=logging.INFO, package=__name__)
-    __slots__ = ("name", "plurals", "extras")
+    __slots__ = ("name", "plurals", "parse_args", "parse_args_locked", "extras")
     def __init__(self, name: str) -> None:
         Flag._log.debug(f"Creating flag: '{name}'")
         self.name = name
         self.plurals: set[str] = set()
+        self.parse_args: list[ParseArg] = list()
+        self.parse_args_locked: bool = False # if True, no additional parse args will be added
         self.extras: set[str] = set()
+        # extras are stored as a defaultdict to act as a lazy ordered set
+        # self.extras: collections.defaultdict[str, None] = collections.defaultdict(lambda: None)
+
+    def get_parse_info(self, defaults: dict[str, typing.Any],
+                       target_regex: str = None) -> tuple[re.Pattern, int]:
+        len_ = 0
+        parse_data: list[str] = [target_regex or "(\d+)"]
+        for arg in self.parse_args:
+            # add pretext
+            len_ += len(arg.pretext)
+            parse_data.append(re.escape(arg.pretext))
+
+            # handle if plural flag
+            if arg.key in constants.PLURAL_FLAGS:
+                parse_data.append(f"(?:{plural_flag_to_plural[arg.key]})?")
+                continue
+            
+            # get the corresponding value from defaults
+            try:
+                default = defaults[arg.key]
+            except KeyError as exc:
+                raise exceptions.ParseError(f"Missing default: {arg.key}") from exc
+            
+            # handle if function
+            if inspect.isfunction(default):
+                parse_data.append(".*?")
+                continue
+
+            # handle arg with str(ifyable) default
+            default = str(default)
+            if arg.required:
+                len_ += len(default)
+                parse_data.append(re.escape(default))
+                continue
+            parse_data.append(f"(?:{re.escape(default)})?")
+
+        return re.compile("".join(parse_data).strip("\\ ")), len_
 
     def get_plurals(self) -> dict[str, str]:
         """Get converted plural flags.
@@ -124,7 +181,7 @@ class Flag:
         return kwargs
 
 
-class Flags(list[Flag]):
+class Flags(list[Flag], AllPretty):
     _log = plogging.setup_new("Flags", level=logging.INFO, package=__name__)
     def __contains__(self, other: Union[str, Flag]) -> bool:
         if isinstance(other, Flag):
@@ -166,7 +223,16 @@ class Flags(list[Flag]):
             return default
 
 
-def _add_parented_flag(field_name: str, _format_spec: str, _conversion: str, flags: Flags) -> str:
+def _add_parse_args(literal_text: str, field_name: str, _format_spec: str, _conversion: str,
+                    required: bool, current_base_flag: Flag) -> None:
+    if not current_base_flag.parse_args_locked:
+        current_base_flag.parse_args.append(ParseArg(field_name, literal_text,
+                                                     _default_fmt(field_name, _format_spec,
+                                                                  _conversion), required))
+
+
+def _add_parented_flag(literal_text: str, field_name: str, _format_spec: str, _conversion: str,
+                       flags: Flags, current_base_flag: Union[Flag, None]) -> str:
     left, _, right = field_name.partition(".")
 
     # determine the parent and child flags
@@ -186,6 +252,15 @@ def _add_parented_flag(field_name: str, _format_spec: str, _conversion: str, fla
         flag = Flag(parent)
         flags.append(flag)
     
+    # add parse args
+    if current_base_flag is not None:
+        if flag.name == current_base_flag.name:
+            parse_args_required = True
+        else:
+            parse_args_required = False
+        _add_parse_args(literal_text, child, _format_spec, _conversion, parse_args_required,
+                        current_base_flag)
+        
     # add the info to the flag
     if child in constants.PLURAL_FLAGS:
         flag.plurals.add(child)
@@ -195,24 +270,26 @@ def _add_parented_flag(field_name: str, _format_spec: str, _conversion: str, fla
     return _mangled_fmt(parent, child, _format_spec, _conversion)
 
 
-def _add_plural_flag(field_name: str, _format_spec: str, _conversion: str, flags: Flags) -> str:
-    if not flags:
+def _add_plural_flag(literal_text: str, field_name: str, _format_spec: str, _conversion: str,
+                     current_base_flag: Union[Flag, None]) -> str:
+    if not current_base_flag:
         raise ValueError("Plural flags may not be used before a base flag if a parent is not "
                          "specified. Try specifying a parent or prefixing the flag name with an "
                          "underscore")
-    flag = flags[-1]
-    flag.plurals.add(field_name)
-    return _mangled_fmt(flag.name, field_name, _format_spec, _conversion)
+    current_base_flag.plurals.add(field_name)
+    _add_parse_args(literal_text, field_name, _format_spec, _conversion, False, current_base_flag)
+    return _mangled_fmt(current_base_flag.name, field_name, _format_spec, _conversion)
 
 
-def _add_extra(field_name: str, _format_spec: str, _conversion: str, flags: Flags) -> str:
-    if not flags:
+def _add_extra(literal_text: str, field_name: str, _format_spec: str, _conversion: str,
+               current_base_flag: Union[Flag, None]) -> str:
+    if not current_base_flag:
         raise ValueError("Extra flags may not be used before a base flag if a parent is not "
                          "specified. Try specifying a parent or prefixing the flag name with an "
                          "underscore")
-    flag = flags[-1]
-    flag.extras.add(field_name)
-    return _mangled_fmt(flag.name, field_name, _format_spec, _conversion)
+    current_base_flag.extras.add(field_name)
+    _add_parse_args(literal_text, field_name, _format_spec, _conversion, True, current_base_flag)
+    return _mangled_fmt(current_base_flag.name, field_name, _format_spec, _conversion)
 
 
 def _add_base_flag(field_name: str, _format_spec: str, _conversion: str, flags: Flags) -> str:
@@ -224,10 +301,11 @@ def _add_base_flag(field_name: str, _format_spec: str, _conversion: str, flags: 
 def update_fmt(fmt: types.SupportsBracketFormat) -> tuple[Flags, types.SupportsBracketFormat]:
     data: list[str] = list()
     flags: Flags = Flags()
+    current_base_flag: Flag = None
     for literal_text, field_name, format_spec, conversion in str_formatter.parse(fmt):
         if literal_text:
             data.append(literal_text)
-    
+
         # usually only happens at the very end if there is leftover text
         if field_name is None and format_spec is None and conversion is None:
             continue
@@ -245,20 +323,27 @@ def update_fmt(fmt: types.SupportsBracketFormat) -> tuple[Flags, types.SupportsB
 
         # field specifies a parent
         if "." in field_name:
-            data.append(_add_parented_flag(field_name, _format_spec, _conversion, flags))
+            data.append(_add_parented_flag(literal_text, field_name, _format_spec, _conversion,
+                                           flags, current_base_flag))
             continue
         
         # field is a plural flag
         if field_name in constants.PLURAL_FLAGS:
-            data.append(_add_plural_flag(field_name, _format_spec, _conversion, flags))
+            data.append(_add_plural_flag(literal_text, field_name, _format_spec, _conversion,
+                                         current_base_flag))
             continue
         
         # field is a base flag
         if field_name in constants.BASE_FLAGS:
             data.append(_add_base_flag(field_name, _format_spec, _conversion, flags))
+            if current_base_flag:
+                current_base_flag.parse_args_locked = True
+            current_base_flag = flags.get(field_name)
             continue
+        
 
-        data.append(_add_extra(field_name, _format_spec, _conversion, flags))
+        data.append(_add_extra(literal_text, field_name, _format_spec, _conversion,
+                               current_base_flag))
     
     return flags, "".join(data)